requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.4.2
pandas==2.2.2
python-dotenv==1.0.1